requirements.txt
Flask==3.0.0
ultralytics==8.0.196
pandas==2.1.3
scikit-learn==1.3.2
beautifulsoup4==4.12.2
requests==2.31.0
Pillow==10.1.0
numpy==1.24.3
matplotlib==3.8.2
Werkzeug==3.0.1
thefuzz==0.19.0
python-Levenshtein==0.21.1

